Introduce new base class util.address.Streaming, deprecating Address
