+++ src/main/php/util/address/XmlStream.class.php
@@ -1,7 +1,5 @@
 <?php namespace util\address;
 
-use Iterator;
-
 /**
  * XML stream input
  *
@@ -10,8 +8,5 @@ use Iterator;
  * @deprecated by XmlStreaming
  */
 class XmlStream extends Address {
-
-  /** Iterator implementation */
-  protected function iterator(): Iterator { return new XmlIterator($this->stream); }
 
 }
